+++ spec/lumberjack/device/log_file_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe Lumberjack::Device::LogFile do
     end
 
     device = Lumberjack::Device::LogFile.new(log_file, template: ":message")
-    device.write(Lumberjack::LogEntry.new(Time.now, 1, "New log entry", nil, $$, nil))
+    device.write(Lumberjack::LogEntry.new(Time.now, 1, "New log entry", nil, Process.pid, nil))
     device.close
 
     expect(File.read(log_file)).to eq("Existing contents\nNew log entry#{Lumberjack::LINE_SEPARATOR}")
@@ -31,11 +31,11 @@ RSpec.describe Lumberjack::Device::LogFile do
     device = Lumberjack::Device::LogFile.new(log_file, keep: 2, buffer_size: 32767)
 
     message = [0xC4, 0x90, 0xE1, 0xBB].pack("c*").force_encoding "ASCII-8BIT"
-    entry = Lumberjack::LogEntry.new(Time.now, 1, message, nil, $$, nil)
+    entry = Lumberjack::LogEntry.new(Time.now, 1, message, nil, Process.pid, nil)
     device.write entry
 
     message = "проверка"
-    entry = Lumberjack::LogEntry.new(Time.now, 1, message, nil, $$, nil)
+    entry = Lumberjack::LogEntry.new(Time.now, 1, message, nil, Process.pid, nil)
     device.write entry
 
     expect do
@@ -46,10 +46,10 @@ RSpec.describe Lumberjack::Device::LogFile do
   it "should reopen the file" do
     log_file = File.join(tmp_dir, "a#{rand(1000000000)}.log")
     device = Lumberjack::Device::LogFile.new(log_file, template: ":message")
-    device.write(Lumberjack::LogEntry.new(Time.now, 1, "message 1", nil, $$, nil))
+    device.write(Lumberjack::LogEntry.new(Time.now, 1, "message 1", nil, Process.pid, nil))
     device.close
     device.reopen
-    device.write(Lumberjack::LogEntry.new(Time.now, 1, "message 2", nil, $$, nil))
+    device.write(Lumberjack::LogEntry.new(Time.now, 1, "message 2", nil, Process.pid, nil))
     device.close
     expect(File.read(log_file)).to eq("message 1#{Lumberjack::LINE_SEPARATOR}message 2#{Lumberjack::LINE_SEPARATOR}")
   end